requirements.txt
llama-index
python-dotenv
streamlit